Agentic-Ai-Clinical-Documentation-Assist/backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-groq
python-dotenv
requests>=2.31.0
streamlit
